chore: extracting voter set
